+++ apps/builder/src/page/App/components/PanelSetters/TableSetter/ColumnSetter/body.tsx
