+++ apps/builder/src/page/App/components/PanelSetters/TableSetter/ColumnSetter/utils/generateNewColumns.ts
