+++ apps/builder/src/widgetLibrary/TableWidget/table.tsx
@@ -5,7 +5,8 @@ import {
   TableWidgetProps,
   WrappedTableProps,
 } from "./interface"
-import { transDataForType } from "@/widgetLibrary/TableWidget/utils"
+import { tansTableDataToColumns, transDataForType } from "@/widgetLibrary/TableWidget/utils"
+import { isEqual } from "lodash"
 
 export const WrappedTable = forwardRef<HTMLInputElement, WrappedTableProps>(
   (props, ref) => {
@@ -90,11 +91,26 @@ export const TableWidget: FC<TableWidgetProps> = (props) => {
     })
     return res
   }, [columns])
+  const isMount = useRef(false)
+
+  useEffect(() => {
+    if (!isMount.current) {
+      isMount.current = true
+      return
+    }
+    if (data?.length) {
+      let newColumns = tansTableDataToColumns(data)
+      if (!isEqual(columns, newColumns)) {
+        handleUpdateDsl({columns: newColumns})
+      }
+    }
+  }, [data])
 
   useEffect(() => {
     handleUpdateGlobalData(displayName, {
       defaultSort,
       data,
+      columns,
     })
     return () => {
       handleDeleteGlobalData(displayName)
@@ -103,6 +119,7 @@ export const TableWidget: FC<TableWidgetProps> = (props) => {
     displayName,
     defaultSort,
     data,
+    columns,
     handleUpdateGlobalData,
     handleUpdateDsl,
     handleDeleteGlobalData,
